UPDATE ::: Removing form and grabbing parameters from request object
git-svn-id: svn://107.170.10.188:8080/Monolith/dev@50306 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 71a3cf186afb45a23916747407f4c2ee23536796

+++ src/prerna/semoss/web/services/local/auth/UserAuthorizationResource.java
@@ -150,7 +150,7 @@ public class UserAuthorizationResource extends AbstractAdminResource {
 	@POST
 	@Produces("application/json")
 	@Path("deleteUserAccessKey")
-	public Response deleteUserAccessKey(@Context HttpServletRequest request, MultivaluedMap<String, String> form) {
+	public Response deleteUserAccessKey(@Context HttpServletRequest request) {
 		User user = null;
 		try {
 			user = ResourceUtility.getUser(request);
@@ -164,7 +164,11 @@ public class UserAuthorizationResource extends AbstractAdminResource {
 		Map<String, Object> retMap = new HashMap<>();
 		
 		AccessToken token = user.getPrimaryLoginToken();
-		String accessKey = form.getFirst("accessKey");
+		String accessKey = request.getParameter("accessKey");
+		if(accessKey == null || accessKey.isEmpty()) {
+			retMap.put(Constants.ERROR_MESSAGE, "accessKey parameter is not defined");
+			return WebUtility.getResponse(retMap, 400);
+		}
 		try {
 			boolean success = SecurityUserAccessKeyUtils.deleteUserAccessToken(token, accessKey);
 			retMap.put("success", success);
